+++ CODEOWNERS
@@ -408,7 +408,8 @@ Makefile* @cilium/build
 /pkg/datapath/linux/config/ @cilium/loader
 /pkg/datapath/linux/ipsec/ @cilium/ipsec
 /pkg/datapath/linux/ipsec/xfrm_collector* @cilium/ipsec @cilium/metrics
-/pkg/datapath/linux/node.go @cilium/ipsec @cilium/sig-datapath
+/pkg/datapath/linux/ipsec.go @cilium/ipsec 
+/pkg/datapath/linux/node.go @cilium/sig-datapath
 /pkg/datapath/linux/probes/ @cilium/loader
 /pkg/datapath/linux/requirements.go @cilium/loader
 /pkg/datapath/types/loader.go @cilium/loader
